# openapi.yml
openapi: 3.0.3
info:
  title: Corazón Migrante API
  version: 1.0.0
  description: "API pública y administrativa de Corazón Migrante.\n\nPrefijo global: `/api/v1`.\n\nContrato de respuesta:\n\
    - Respuesta simple: `{ data, meta }`.\n- Respuesta paginada: `{ data, pagination, meta }`.\n- Error: `{ error: { code,\
    \ message, details }, meta }`.\n\nSeguridad:\n- Los endpoints públicos están marcados como `public` en su descripción.\n\
    - Los endpoints privados usan `Bearer JWT`.\n- Algunos endpoints administrativos requieren permisos específicos además\
    \ del rol.\n\nDisponibilidad de terapeutas:\n- `GET /booking/availability` calcula los horarios disponibles por terapeuta.\n\
    - Parte de los horarios recurrentes configurados en `therapist_schedules`.\n- Resta bloqueos activos de agenda (`therapist_blocked_times`).\n\
    - Resta citas activas/ocupadas (`appointments` con status `REQUESTED` o `CONFIRMED`).\n- El rango máximo soportado es\
    \ 31 días.\n\nLímites de uso (rate limiting):\n- Límite global por IP: 120 solicitudes / 60s. Al excederlo, la API responde\
    \ `429`\n  con `{ error: { code: \"HTTP_429\", ... } }`.\n- `POST /auth/login`, `/auth/register/*`: 10 solicitudes / 60s\
    \ por IP.\n- `POST /auth/password-reset/request` y `/confirm`: 5 solicitudes / 60s por IP.\n- El cuerpo de las solicitudes\
    \ JSON está limitado a 2 MB (`413` si se excede).\n"
servers:
- url: https://api.corazondemigrante.com/api/v1
  description: Producción
...
security:
- bearerAuth: []
paths:
...
  /auth/register/patient:
    post:
      tags:
      - Auth
      summary: Registrar paciente
      description: Público. Crea usuario con rol de paciente y su perfil inicial.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterPatientRequest'
      responses:
        '201':
          $ref: '#/components/responses/CreatedEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '429':
          $ref: '#/components/responses/TooManyRequestsError'
  /auth/register/therapist:
    post:
      tags:
      - Auth
      summary: Registrar terapeuta
      description: Público. Crea usuario terapeuta y perfil profesional inicial.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterTherapistRequest'
      responses:
        '201':
          $ref: '#/components/responses/CreatedEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '429':
          $ref: '#/components/responses/TooManyRequestsError'
  /auth/login:
    post:
      tags:
      - Auth
      summary: Iniciar sesión
      description: Público. Devuelve tokens de acceso y refresh token.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: paciente.demo@corazonmigrante.test
              password: Demo123456!
      responses:
        '200':
          description: Sesión iniciada.
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/EnvelopeObject'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/AuthSession'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/TooManyRequestsError'
  /auth/refresh:
    post:
      tags:
      - Auth
      summary: Renovar access token
      description: Público. Usa un refresh token válido para emitir una nueva sesión.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          $ref: '#/components/responses/OkEnvelope'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /auth/logout:
    post:
      tags:
      - Auth
      summary: Cerrar sesión
      description: Público. Invalida el refresh token enviado.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          $ref: '#/components/responses/OkEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
  /auth/password-reset/request:
    post:
      tags:
      - Auth
      summary: Solicitar recuperación de contraseña
      description: Público. Encola o envía instrucciones de recuperación si el correo existe.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
      responses:
        '200':
          $ref: '#/components/responses/OkEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequestsError'
  /auth/password-reset/confirm:
    post:
      tags:
      - Auth
      summary: Confirmar recuperación de contraseña
      description: Público. Cambia contraseña usando token de recuperación válido.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetConfirmRequest'
      responses:
        '200':
          $ref: '#/components/responses/OkEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/TooManyRequestsError'
...
  /appointments:
    post:
      tags:
      - Appointments
      summary: Crear solicitud de cita
      description: 'Requiere rol `PATIENT`.

        Antes de crear la cita, valida que el slot no esté ocupado por otra cita activa ni bloqueado por agenda.

        '
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAppointmentRequest'
      responses:
        '201':
          $ref: '#/components/responses/CreatedEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /appointments/mine:
    get:
      tags:
      - Appointments
      summary: Listar mis citas
      description: 'Requiere autenticación.

        Devuelve las citas del usuario autenticado según su rol.

        Acepta paginación canónica `page`/`limit` y aliases legacy `p_page`/`p_limit`.

        '
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/PPage'
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/PLimit'
      - $ref: '#/components/parameters/PageSize'
      - $ref: '#/components/parameters/Sort'
      - $ref: '#/components/parameters/SortBy'
      - $ref: '#/components/parameters/Order'
      - $ref: '#/components/parameters/SortDir'
      - $ref: '#/components/parameters/Search'
      - $ref: '#/components/parameters/Q'
      responses:
        '200':
          $ref: '#/components/responses/PaginatedEnvelope'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /appointments/{id}/status:
    patch:
      tags:
      - Appointments
      summary: Cambiar estado de una cita
      description: Paciente, terapeuta o admin pueden cambiar estado según reglas del servicio.
      parameters:
      - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAppointmentStatusRequest'
      responses:
        '200':
          $ref: '#/components/responses/OkEnvelope'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /appointments/admin/list:
    get:
      tags:
      - Appointments
      summary: Listar citas desde administración
      description: Requiere rol `ADMIN` o `SUPER_ADMIN` y permiso `appointments:read`.
      parameters:
      - $ref: '#/components/parameters/Page'
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Search'
      - $ref: '#/components/parameters/Q'
      responses:
        '200':
          $ref: '#/components/responses/PaginatedEnvelope'
...
  /public/pages/{slug}:
    get:
      tags:
      - Public CMS
      summary: Obtener página pública CMS por slug
      description: Público. Ruta canónica para páginas públicas configurables.
      security: []
      parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
          example: inicio
      responses:
        '200':
          $ref: '#/components/responses/OkEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /admin/public-pages:
    get:
      tags:
      - Admin CMS
      summary: Listar páginas públicas administrables
      description: Alias de negocio para no depender del nombre técnico CMS en clientes nuevos.
      parameters:
      - name: status
        in: query
        schema:
          type: string
          enum:
          - DRAFT
          - PUBLISHED
      responses:
        '200':
          $ref: '#/components/responses/OkEnvelope'
    post:
      tags:
      - Admin CMS
      summary: Crear página pública administrable
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCmsPageRequest'
      responses:
        '201':
          $ref: '#/components/responses/CreatedEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
  /admin/public-pages/{pageId}:
    get:
      tags:
      - Admin CMS
      summary: Obtener página pública administrable
      parameters:
      - name: pageId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          $ref: '#/components/responses/OkEnvelope'
        '404':
          $ref: '#/components/responses/NotFoundError'
    patch:
      tags:
      - Admin CMS
      summary: Actualizar página pública administrable
      parameters:
      - name: pageId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCmsPageRequest'
      responses:
        '200':
          $ref: '#/components/responses/OkEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    delete:
      tags:
      - Admin CMS
      summary: Eliminar página pública administrable
      parameters:
      - name: pageId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          $ref: '#/components/responses/OkEnvelope'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /admin/cms/pages:
    post:
      tags:
      - Admin CMS
      summary: Crear página CMS
      description: Requiere rol admin y permiso `cms:write`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCmsPageRequest'
      responses:
        '201':
          $ref: '#/components/responses/CreatedEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /admin/cms/pages/{pageId}/elements:
    post:
      tags:
      - Admin CMS
      summary: Crear elemento de página CMS
      parameters:
      - name: pageId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCmsElementRequest'
      responses:
        '201':
          $ref: '#/components/responses/CreatedEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
...
  /api/v1/admin/public-pages/{pageId}/posts:
    post:
      tags:
      - Páginas públicas admin
      summary: Asocia una publicación a una página pública dinámica
      security: *id001
      parameters:
      - name: pageId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - publicationId
              properties:
                publicationId:
                  type: string
                  format: uuid
      responses:
        '201':
          description: Publicación asociada a la página pública.
          content: *id002
  /api/v1/admin/public-pages/{pageId}/posts/{postId}:
    delete:
      tags:
      - Páginas públicas admin
      summary: Desasocia una publicación de una página pública dinámica
      security: *id001
      parameters:
      - name: pageId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: postId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Publicación desasociada de la página pública.
          content: *id002
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Página actual. Tiene prioridad sobre `p_page`.
    PPage:
      name: p_page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
      description: Alias legacy aceptado para paginación. `page` tiene prioridad cuando ambos llegan.
    PageSize:
      name: pageSize
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Tamaño de página histórico.
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Tamaño de página recomendado para frontend/smoke. Tiene prioridad sobre `pageSize` y `p_limit`.
    PLimit:
      name: p_limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
      description: Alias legacy aceptado para tamaño de página. `limit` y `pageSize` tienen prioridad cuando llegan juntos.
    Search:
      name: search
      in: query
      required: false
      schema:
        type: string
      description: Búsqueda textual canónica.
    Q:
      name: q
      in: query
      required: false
      schema:
        type: string
      description: Alias corto para búsqueda. `search` tiene prioridad cuando ambos llegan.
    Sort:
      name: sort
      in: query
      required: false
      schema:
        type: string
        default: createdAt
      description: Campo de orden. El backend normaliza alias camelCase/snake_case seguros y evita ORDER BY con columnas inexistentes.
    SortBy:
      name: sortBy
      in: query
      required: false
      schema:
        type: string
      description: Alias frontend para `sort`. `sort` tiene prioridad si ambos llegan.
    Order:
      name: order
      in: query
      required: false
      schema:
        type: string
        enum:
        - asc
        - desc
        default: desc
    SortDir:
      name: sortDir
      in: query
      required: false
      schema:
        type: string
        enum:
        - asc
        - desc
        default: desc
      description: Alias frontend para `order`. `order` tiene prioridad si ambos llegan.
...
  responses:
    OkEnvelope:
      description: Respuesta exitosa envuelta en `{ data, meta }`.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/EnvelopeObject'
    CreatedEnvelope:
      description: Recurso creado. Respuesta envuelta en `{ data, meta }`.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/EnvelopeObject'
    PaginatedEnvelope:
      description: Lista paginada envuelta en `{ data, pagination, meta }`.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PaginatedEnvelope'
    ValidationError:
      description: Error de validación o contrato de entrada.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: HTTP_400
              message:
              - property debug should not exist
              details:
              - property debug should not exist
            meta:
              requestId: 315c2096-df71-447a-92ad-2f9329316896
              timestamp: '2026-07-05T15:13:13.536Z'
    UnauthorizedError:
      description: Token ausente, vencido o inválido.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    ForbiddenError:
      description: Usuario autenticado sin rol o permiso suficiente.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFoundError:
      description: Recurso no encontrado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    ConflictError:
      description: Conflicto de unicidad, solapamiento o regla de negocio.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    TooManyRequestsError:
      description: Límite de solicitudes excedido (rate limiting).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: HTTP_429
              message: 'ThrottlerException: Too Many Requests'
              details: []
            meta:
              requestId: 315c2096-df71-447a-92ad-2f9329316896
              timestamp: '2026-07-05T15:13:13.536Z'
  schemas:
    Meta:
      type: object
      required:
      - requestId
      - timestamp
      properties:
        requestId:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
    EnvelopeObject:
      type: object
      required:
      - data
      - meta
      properties:
        data:
          oneOf:
          - type: object
            nullable: true
          - type: array
            items: {}
          - type: string
          - type: boolean
          - type: number
        meta:
          $ref: '#/components/schemas/Meta'
    Pagination:
      type: object
      required:
      - page
      - pageSize
      - limit
      - total
      - totalPages
      properties:
        page:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 20
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 138
        totalPages:
          type: integer
          example: 7
    PaginatedEnvelope:
      type: object
      required:
      - data
      - pagination
      - meta
      properties:
        data:
          type: array
          items: {}
        pagination:
          $ref: '#/components/schemas/Pagination'
        meta:
          $ref: '#/components/schemas/Meta'
    ErrorEnvelope:
      type: object
      required:
      - error
      - meta
      properties:
        error:
          type: object
          required:
          - code
          - message
          - details
          properties:
            code:
              type: string
              example: HTTP_400
            message:
              oneOf:
              - type: string
              - type: array
                items:
                  type: string
            details:
              type: array
              items: {}
        meta:
          $ref: '#/components/schemas/Meta'
    LoginRequest:
      type: object
      required:
      - email
      - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
    RefreshTokenRequest:
      type: object
      required:
      - refreshToken
      properties:
        refreshToken:
          type: string
    PasswordResetRequest:
      type: object
      required:
      - email
      properties:
        email:
          type: string
          format: email
    PasswordResetConfirmRequest:
      type: object
      required:
      - email
      - pin
      - newPassword
      properties:
        email:
          type: string
          format: email
        pin:
          type: string
          minLength: 6
          maxLength: 6
          example: '123456'
        newPassword:
          type: string
          format: password
          minLength: 8
    AuthSession:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          type: object
          properties:
            id:
              type: string
              format: uuid
            email:
              type: string
              format: email
            role:
              type: string
            permissions:
              type: array
              items:
                type: string
    RegisterPatientRequest:
      type: object
      required:
      - email
      - password
      - firstName
      - lastName
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        birthDate:
          type: string
          format: date
        country:
          type: string
        city:
          type: string
        occupation:
          type: string
    RegisterTherapistRequest:
      type: object
      required:
      - email
      - password
      - firstName
      - lastName
      - title
      - mainSpecialty
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        title:
          type: string
        mainSpecialty:
          type: string
        bio:
          type: string
        personalPhrase:
          type: string
        youtubeUrl:
          type: string
          format: uri
        licenseNumber:
          type: string
        country:
          type: string
        city:
          type: string
        baseSessionPrice:
          type: number
          minimum: 0
...
    CreateAppointmentRequest:
      type: object
      required:
      - therapistUserId
      - productId
      - scheduledStartAt
      - timezone
      properties:
        therapistUserId:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        scheduledStartAt:
          type: string
          format: date-time
        timezone:
          type: string
          example: America/La_Paz
        notesForTherapist:
          type: string
    UpdateAppointmentStatusRequest:
      type: object
      required:
      - status
      properties:
        status:
          type: string
          enum:
          - CONFIRMED
          - COMPLETED
          - CANCELLED_BY_PATIENT
          - CANCELLED_BY_ADMIN
          - CANCELLED_BY_THERAPIST
          - NO_SHOW
        reason:
          type: string
    CreateCmsPageRequest:
      type: object
      required:
      - slug
      - title
      properties:
        slug:
          type: string
          example: inicio
        title:
          type: string
        status:
          type: string
          enum:
          - DRAFT
          - PUBLISHED
        seoMetadata:
          type: object
          additionalProperties: true
    CreateCmsElementRequest:
      type: object
      required:
      - code
      - type
      - content
      properties:
        code:
          type: string
          example: hero
        type:
          type: string
          example: JSON
        content:
          type: object
          additionalProperties: true
        sortOrder:
          type: integer
...